third_party/blink/web_tests/external/wpt/html/dom/documents/resource-metadata-management/document-readyState-expected.txt

This is a testharness.js-based test.
[FAIL] readyState equals 'complete' for documents created via createHTMLDocument()
  assert_equals: expected "complete" but got "loading"
Harness: the test ran to completion.
